qcacld-3.0: Don't set TDLS mode on inactive STA vdev
When scan is done, we are setting the TDLS mode on
STA vdev, if Sta vdev is not there then we are enabling
on P2P CLI vdev. This causes the wrong vdev 0 which is
available for scan to be fetched as the TDLS vdev.
So the peer update timer is not restarted on P2P
CLI vdev, this causes the peer to be retained after
scan even if traffic is stopped.

Check if the vdev is connected to fetch the TDLS vdev.

Change-Id: If058e8c29141129549a9c688d013290c44fe2000
CRs-Fixed: 3800379